Fix setter_user_id comparisons in doors_subquery: cast to int for type safety

+++ app/api/repcard/unified-dashboard/route.ts
@@ -519,14 +519,14 @@ export async function GET(request: NextRequest) {
                     MIN(c2.created_at) as day_min,
                     MAX(c2.created_at) as day_max
                   FROM repcard_customers c2
-                  WHERE c2.setter_user_id = u.repcard_user_id
+                  WHERE c2.setter_user_id::int = u.repcard_user_id
                     AND c2.created_at >= ${startDate}::timestamptz
                     AND c2.created_at <= ${endDate}::timestamptz
                   GROUP BY DATE(c2.created_at AT TIME ZONE 'America/New_York')
                 ) daily_ranges
               ), 0)::int as estimated_hours_on_doors
             FROM repcard_customers c
-            WHERE c.setter_user_id = u.repcard_user_id
+            WHERE c.setter_user_id::int = u.repcard_user_id
               AND c.created_at >= ${startDate}::timestamptz
               AND c.created_at <= ${endDate}::timestamptz
           ) doors_subquery ON true
@@ -571,12 +571,12 @@ export async function GET(request: NextRequest) {
                     MIN(c2.created_at) as day_min,
                     MAX(c2.created_at) as day_max
                   FROM repcard_customers c2
-                  WHERE c2.setter_user_id = u.repcard_user_id
+                  WHERE c2.setter_user_id::int = u.repcard_user_id
                   GROUP BY DATE(c2.created_at AT TIME ZONE 'America/New_York')
                 ) daily_ranges
               ), 0)::int as estimated_hours_on_doors
             FROM repcard_customers c
-            WHERE c.setter_user_id = u.repcard_user_id
+            WHERE c.setter_user_id::int = u.repcard_user_id
           ) doors_subquery ON true
           WHERE u.repcard_user_id IS NOT NULL
           GROUP BY u.repcard_user_id, u.name, u.role, doors_subquery.doors_knocked, doors_subquery.estimated_hours_on_doors
